Un-skip and properly nest tests for empty body and error cases
Left a couple of the empty body related tests marked as skip because
they are asserting that a __typename will be supplied even if the body
is empty. The tests don't supply an empty body, so they don't seem
valid. And it doesn't seem like we would be able to generate the
__typename with no type data (no type field in the body). Not sure how
we want to handle those cases yet.